move nonescaping check to TypeDiagnostics
